+++ R/data-converters.R
@@ -264,7 +264,8 @@ enw_incidence_to_linelist <- function(obs, reference_date = "reference_date",
 #' concerned about runtime. Note that the start of the timestep will be
 #' determined by `min_date` + a single timestep (i.e. the
 #' first timestep will be "2022-10-23" if the minimum reference date is
-#' "2022-10-16").
+#' "2022-10-16"). Observations where the report dates do not form a complete
+#' timestep will be dropped from the aggregated output.
 #'
 #' @param obs An object coercible to a `data.table` (such as a `data.frame`)
 #' which must have a `new_confirm` numeric column, and `report_date` and
@@ -308,6 +309,9 @@ enw_aggregate_cumulative <- function(
     required_cols = c("confirm", by), forbidden_cols = ".group",
     dates = TRUE, copy = copy
   )
+  if (nrow(obs) < 2) {
+    cli::cli_abort("There must be at least two observations")
+  }
   obs <- enw_complete_dates(obs, by = by, timestep = "day")
   obs <- enw_assign_group(obs, by = by)
   check_timestep_by_date(obs, timestep = "day", exact = TRUE)